# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enerji Yönetim Sistemi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: Tüm birimleri listele
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Yeni birim ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Birim detayı
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Birim güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Birim sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── METERS ──────────────────────────────────────────────
...
  /swot:
    get:
      operationId: listSwotItems
      tags: [swot]
      summary: SWOT maddeleri listesi
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SwotItem"
    post:
      operationId: createSwotItem
      tags: [swot]
      summary: SWOT maddesi ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwotItem"
  /swot/{id}:
    patch:
      operationId: updateSwotItem
      tags: [swot]
      summary: SWOT maddesi güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwotUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwotItem"
    delete:
      operationId: deleteSwotItem
      tags: [swot]
      summary: SWOT maddesi sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── TARGETS ───────────────────────────────────────────────
  /targets:
    get:
      operationId: listTargets
      tags: [targets]
      summary: Enerji hedeflerini listele (ilerleme dahil)
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnergyTargetWithProgress"
    post:
      operationId: createTarget
      tags: [targets]
      summary: Enerji hedefi ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnergyTargetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnergyTarget"
  /targets/{id}:
    patch:
      operationId: updateTarget
      tags: [targets]
      summary: Enerji hedefini güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnergyTargetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnergyTarget"
    delete:
      operationId: deleteTarget
      tags: [targets]
      summary: Enerji hedefini sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── RISKS ─────────────────────────────────────────────────
...
  /seu:
    get:
      operationId: listSeu
      tags: [seu]
      summary: Önemli enerji kullanımları listesi
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeuItem"
    post:
      operationId: createSeu
      tags: [seu]
      summary: ÖEK ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeuInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeuItem"
  /seu/{id}:
    patch:
      operationId: updateSeu
      tags: [seu]
      summary: ÖEK güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeuUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeuItem"
    delete:
      operationId: deleteSeu
      tags: [seu]
      summary: ÖEK sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────
...
components:
  schemas:
...
    Unit:
      type: object
      required: [id, name, location, type, city, active]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        type:
          type: string
          description: "fabrika | ofis | depo | hastane | okul | diger"
        city:
          type: string
        responsible:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    UnitInput:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          type: string
        type:
          type: string
        city:
          type: string
        responsible:
          type: string
        description:
          type: string
        active:
          type: boolean
    UnitUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        type:
          type: string
        city:
          type: string
        responsible:
          type: string
        description:
          type: string
        active:
          type: boolean

    # ── Meter ───────────────────────────────────────────────
...
    SwotItem:
      type: object
      required: [id, category, title, score, impact]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        category:
          type: string
          description: "strengths | weaknesses | opportunities | threats"
        title:
          type: string
        description:
          type: ["string", "null"]
        score:
          type: integer
          description: "1-5 puan"
        impact:
          type: string
          description: "yuksek | orta | dusuk"
        createdAt:
          type: string
    SwotInput:
      type: object
      required: [category, title, score, impact]
      properties:
        unitId:
          type: integer
        category:
          type: string
        title:
          type: string
        description:
          type: string
        score:
          type: integer
        impact:
          type: string
    SwotUpdate:
      type: object
      properties:
        unitId:
          type: integer
        category:
          type: string
        title:
          type: string
        description:
          type: string
        score:
          type: integer
        impact:
          type: string

    # ── Risk ─────────────────────────────────────────────────
...
    SeuItem:
      type: object
      required: [id, name, category, annualKwh, percentage, priority]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        annualKwh:
          type: number
        percentage:
          type: number
        priority:
          type: integer
        targetReductionPercent:
          type: ["number", "null"]
        responsible:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SeuInput:
      type: object
      required: [name, category]
      properties:
        unitId:
          type: integer
        name:
          type: string
        category:
          type: string
        annualKwh:
          type: number
        percentage:
          type: number
        priority:
          type: integer
        targetReductionPercent:
          type: number
        responsible:
          type: string
        notes:
          type: string
    SeuUpdate:
      type: object
      properties:
        unitId:
          type: integer
        name:
          type: string
        category:
          type: string
        annualKwh:
          type: number
        percentage:
          type: number
        priority:
          type: integer
        targetReductionPercent:
          type: number
        responsible:
          type: string
        notes:
          type: string

    # ── Dashboard ─────────────────────────────────────────────
...
    EnergyTarget:
      type: object
      required: [id, name, baselineYear, targetYear, targetReductionPercent, createdAt]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    EnergyTargetWithProgress:
      type: object
      required: [id, name, baselineYear, targetYear, targetReductionPercent, createdAt, yearlyProgress]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        baselineKwh:
          type: ["number", "null"]
        yearlyProgress:
          type: array
          items:
            type: object
            required: [year]
            properties:
              year:
                type: integer
              actualKwh:
                type: ["number", "null"]
              reductionPercent:
                type: ["number", "null"]
    EnergyTargetInput:
      type: object
      required: [name, baselineYear, targetYear, targetReductionPercent]
      properties:
        unitId:
          type: integer
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: string
    EnergyTargetUpdate:
      type: object
      properties:
        unitId:
          type: integer
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: string
...